backend/requirements.txt
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.1.1
uvicorn[standard]==0.34.0
